fix: show actual V1 usage above plan limits (#4199)
* fix: show over-limit usage in V1 non-interactive mode

* docs: describe usage fix as Classic CLI

* fix: address V1 usage review feedback

* Update .changes/20260819-1554-fixed-classic-cli-usage-over-limit.json


* fix: align usage percentage presentation

---------

+++ crates/chat-cli/src/cli/chat/cli/usage/mod.rs
@@ -48,7 +48,7 @@ pub struct UsageBreakdownInfo {
     pub display_name: String,
     pub used: f64,
     pub limit: f64,
-    pub percentage: i32,
+    pub percentage: f64,
     /// Legacy post-paid overage fields; retained from the API but no longer rendered.
     #[allow(dead_code)]
     pub current_overages: f64,
